services/chat-service/requirements.txt
fastapi
uvicorn
boto3
kafka-python
uuid
